Layout and components completed.
