+++ frontend/src/components/index.js
@@ -1,0 +1,5 @@
+import AdditionalDetails from './AdditionalDetails'
+import FilterBar from './FilterBar'
+import TripList from './TripList'
+
+export {AdditionalDetails, FilterBar, TripList}
